refactor(claim-write): create claim_ingest app; relocate the bulk pipeline
Move the bulk ingest pipeline out of catalog into its own top-level app,
apps.claim_ingest — pure code motion now that Steps 1-2 severed its catalog
couplings (the LinkableClaimModel bound and the resolution dispatch). The app
owns no concrete models, so no migration; IngestRun stays in provenance.

- git mv apps/catalog/ingestion -> apps/claim_ingest (patches/, plan.py, apply/,
  r2_pull.py ride along) + trivial AppConfig + INSTALLED_APPS entry.
- Rewrite the package's absolute self-imports and every external importer from
  apps.catalog.ingestion -> apps.claim_ingest (test files, citation/seeding
  docstring, the test_patches mock-target string). ruff re-sorted the touched
  import blocks.
- ingest_patches / pull_patches / dump_patch_entry stay in catalog/management:
  ingest_patches composes the pipeline with catalog cache invalidation, so it is
  the orchestration entrypoint (apps.catalog.management is outside both ingest
  contracts' source lists) — keeping it there keeps claim_ingest free of catalog.
- Tests stay in catalog/tests (repointed): they are catalog-integration tests
  bound to catalog/tests/conftest.py fixtures, not portable engine tests.

Contracts (pyproject.toml):
- Retarget the two ingest forbids and the patch front/back-end split onto
  apps.claim_ingest. The forward forbid is now a single whole-app
  "apps.claim_ingest -/-> apps.catalog" (claim_ingest is no longer a descendant
  of catalog, so the source need not be enumerated; it subsumes catalog.api).
- Drop the now-stale apps.catalog.ingestion entry from the api-leaf contract —
  the whole-app forward forbid already covers claim_ingest -> catalog.api.
- Add apps.claim_ingest to the "Data apps do not depend on media" wall.

Docs: update AppBoundaries.md to the implemented contracts; record the Step 3
deviations and Step 4 refinements (relative->absolute repoint, tests stay put,
StructuredValidationError test importers) in the plan.

lint-imports 14/14, mypy clean, full backend suite green.
