+++ backend/apps/catalog/tests/test_apply.py
@@ -8,15 +8,15 @@ import pytest
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
 
-from apps.catalog.ingestion.apply import apply_plan
-from apps.catalog.ingestion.plan import (
+from apps.catalog.models import CorporateEntity, Manufacturer, Theme
+from apps.claim_ingest.apply import apply_plan
+from apps.claim_ingest.plan import (
     IngestPlan,
     PlannedClaimAssert,
     PlannedClaimRetract,
     PlannedEntityCreate,
     RunReport,
 )
-from apps.catalog.models import CorporateEntity, Manufacturer, Theme
 from apps.provenance.models import ChangeSet, Claim, IngestRun, Source
 
 pytestmark = pytest.mark.django_db
