+++ backend/apps/catalog/tests/test_dump_patch_entry.py
@@ -17,11 +17,11 @@ from django.contrib.auth import get_user_model
 from django.core.management import call_command
 from django.core.management.base import CommandError
 
-from apps.catalog.ingestion.apply import apply_plan
-from apps.catalog.ingestion.patches import EditEntry, build_plan, load_patch
 from apps.catalog.models import MachineModel
 from apps.catalog.tests.conftest import make_machine_model
 from apps.citation.models import CitationSource
+from apps.claim_ingest.apply import apply_plan
+from apps.claim_ingest.patches import EditEntry, build_plan, load_patch
 from apps.core.markdown import convert_authoring_to_storage
 from apps.provenance.models import CitationInstance, Claim, Source
 from apps.provenance.test_factories import user_changeset
