+++ backend/apps/catalog/tests/test_patch_provenance.py
@@ -12,21 +12,21 @@ import pytest
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
 
-from apps.catalog.ingestion.apply import apply_plan
-from apps.catalog.ingestion.patches import (
+from apps.catalog.models import MachineModel, Manufacturer, Tag
+from apps.catalog.tests.conftest import make_machine_model
+from apps.citation.models import CitationSource
+from apps.claim_ingest.apply import apply_plan
+from apps.claim_ingest.patches import (
     EditEntry,
     PatchError,
     build_plan,
     load_patch,
 )
-from apps.catalog.ingestion.plan import (
+from apps.claim_ingest.plan import (
     CitationRef,
     IngestPlan,
     PlannedClaimAssert,
 )
-from apps.catalog.models import MachineModel, Manufacturer, Tag
-from apps.catalog.tests.conftest import make_machine_model
-from apps.citation.models import CitationSource
 from apps.provenance.models import (
     ChangeSet,
     CitationInstance,
