+++ backend/apps/claim_ingest/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ClaimIngestConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.claim_ingest"
+    verbose_name = "Claim ingest"
