+++ backend/config/settings.py
@@ -52,6 +52,7 @@ INSTALLED_APPS = [
     "apps.accounts",
     "apps.core",
     "apps.catalog",
+    "apps.claim_ingest",
     "apps.citation",
     "apps.provenance",
     "apps.media",
